Added Signup Page
Added a signup page that would read input from users to create a new user. This new user will later have their information stored in a database where it can be pulled up after they log in. Also made slight changes to the navbar so that the title is centered.